+++ docs/source/history.md
@@ -411,12 +411,18 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
 
 ## v0.5.3
 
+- Core:
+  - This version introduces `InductionModel` classes, computing the induction factor `a(ct)`
 - Wake models:
-  - Removing `sbeta` limitation from `Bastankhah2014` and `CrespoHernandez`
-  - New option `regularize` for `Bastankhah2014`, `TurbOParkWake` and `TurbOParkIXWake`, smoothing the sqrt at the end of the near wake region
+  - Reformulating `beta` in terms of induction in `Bastankhah2014` and `CrespoHernandez`
+  - New default: `Bastankhah2014` now with default induction model `Madsen`
+  - New default: `TurbOPark` and `TurbOParkIX` now with default induction model `Madsen`
   - Introducing `WakeMirror` wrapper around wake models, modelling wake reflection from ground or horizontal plane via virtual mirrored turbines
 - Wake frames:
   - Renaming `Streamlines` to `Streamlines2D`, no changes in model book names
+- Induction models:
+  - New induction model: `BetzInduction`, the classic `a = 0.5(1 - sqrt(1-ct))` relation
+  - New induction model: `MadsenInduction`, a third-order polynomial approximation of `a(ct)`
 - Bug fixes:
   - Fix for bug in `Streamlines2D` when used in combination with `WakeMirror`
 - Tests: